refactor(models): split structs and enums
