+++ synth-converter/src/models/mod.rs
@@ -1,5 +1,6 @@
-pub mod graph_impls;
 pub mod types;
+pub mod enums;
 
 // Re-export all models;
 pub use types::*;
+pub use enums::*;
