+++ synth-converter/src/models/types.rs
@@ -1,9 +1,27 @@
+
 // The structure follows the input data as descibed in the
 // https://github.com/sdsc-ordes/cat-plus-ontology see here for the expected Synth input data:
 // https://github.com/sdsc-ordes/cat-plus-ontology/tree/96091fd2e75e03de8a4c4d66ad502b2db27998bd/json-file/1-Synth
+use crate::{
+    models::enums::{ActionName, Unit},
+    graph::{
+        insert_into::{InsertIntoGraph, Link},
+        namespaces::{alloproc, alloqual, allores, cat, obo, purl, qudt, schema},
+    },
+};
+use anyhow;
 use serde::{Deserialize, Serialize};
+use sophia::{
+    api::ns::{rdf, xsd},
+    inmem::graph::LightGraph,
+};
+use sophia_api::{
+    graph::MutableGraph,
+    term::{SimpleTerm, Term},
+};
 
-#[derive(Debug, Serialize, Deserialize)]
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Batch {
     #[serde(rename = "batchID")]
@@ -19,7 +37,37 @@ pub struct Batch {
     pub link: Option<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+impl InsertIntoGraph for Batch {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+
+
+        for (pred, value) in [
+            (rdf::type_, &cat::Batch.as_simple()),
+            (schema::name, &self.batch_id.as_simple()),
+        ] {
+            value.attach_into(
+            graph,
+            Link {
+                source_iri: iri.clone(),
+                pred: pred.as_simple(),
+                target_iri: None,
+            },
+            )?;
+        }
+
+        // NOTE: for actions, the direction is reversed (action hasbatch batch)
+        for action in &self.actions {
+
+            let action_uri = action.get_uri();
+            graph.insert(&action_uri, cat::hasBatch.as_simple(), iri.clone())?;
+            action.insert_into(graph, action_uri)?;
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Action {
     pub action_name: ActionName,
@@ -41,18 +89,46 @@ pub struct Action {
     pub pressure_measurement: Option<Observation>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-#[allow(non_snake_case, non_camel_case_types)]
-pub enum ActionName {
-    AddAction,
-    setTemperatureAction,
-    filtrateAction,
-    shakeAction,
-    setVacuumAction,
-    setPressureAction,
+impl InsertIntoGraph for Action {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+
+        for (pred, value) in [
+            (rdf::type_,           &self.action_name.iri().as_simple() as &dyn InsertIntoGraph),
+            (allores::AFX_0000622, &(self.start_time.as_str() * xsd::dateTime).as_simple()),
+            (allores::AFR_0002423, &(self.ending_time.as_str() * xsd::dateTime).as_simple()),
+            (allores::AFR_0001606, &self.method_name.as_simple()),
+            (allores::AFR_0001723, &self.equipment_name.as_simple()),
+            (cat::subEquipmentName,&self.sub_equipment_name.as_simple()),
+            (cat::speedInRPM,      &self.speed_shaker),
+            (cat::temperatureTumbleStirrerShape, &self.temperature_tumble_stirrer),
+            (cat::speedTumbleStirrerShape, &self.speed_tumble_stirrer),
+            (cat::temperatureShakerShape, &self.temperature_shaker),
+            (alloproc::AFP_0002677, &self.pressure_measurement),
+            (cat::hasSample, &self.has_sample),
+            (cat::hasContainerPositionAndQuantity, &self.has_container_position_and_quantity),
+            (alloqual::AFQ_0000111, &self.dispense_state.as_ref().clone().map(|s| s.as_simple())),
+            (cat::dispenseType,    &self.dispense_type.as_ref().clone().map(|s| s.as_simple())),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: pred.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+
+        // NOTE: for container_info, we attach triples directly to the action
+        let _ = &self.container_info.insert_into(graph, iri.clone())?;
+
+        Ok(())
+    }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ContainerInfo {
     #[serde(rename = "containerID")]
@@ -60,21 +136,99 @@ pub struct ContainerInfo {
     pub container_barcode: String,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+impl InsertIntoGraph for ContainerInfo {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+        
+        for (prop, value) in [
+            (&cat::containerID, &self.container_id.as_simple() as &dyn InsertIntoGraph),
+                       
+            (&cat::containerBarcode, &self.container_barcode.as_simple()),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: prop.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Observation {
     pub value: f64,
-    pub unit: String,
+    pub unit: Unit,
     pub error_margin: Option<ErrorMargin>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ErrorMargin {
-    pub value: f64,
-    pub unit: String,
+
+/// Implementation for concrete [Observation].
+impl InsertIntoGraph for Observation {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+        for (prop, value) in [
+            /*
+            (
+                rdf::type_,
+                &cat::Observation.as_simple() as &dyn InsertIntoGraph,
+            ),
+            */
+            (qudt::unit, &self.unit.iri().as_simple() as &dyn InsertIntoGraph),
+            (qudt::value, &self.value.as_simple()),
+            (cat::errorMargin, &self.error_margin),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: prop.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+
+        Ok(())
+    }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ErrorMargin {
+    pub value: f64,
+    pub unit: Unit,
+}
+
+
+/// Implementation for concrete [Observation].
+impl InsertIntoGraph for ErrorMargin {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+        for (prop, value) in [
+            /*
+            (
+                rdf::type_,
+                &cat::errorMargin.as_simple() as &dyn InsertIntoGraph,
+            ),
+            */
+            (qudt::unit, &self.unit.iri().as_simple() as &dyn InsertIntoGraph),
+            (qudt::value, &self.value.as_simple()),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: prop.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Sample {
     #[serde(flatten)]
@@ -87,7 +241,35 @@ pub struct Sample {
     pub has_sample: Vec<SampleItem>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+impl InsertIntoGraph for Sample {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+    
+        for (prop, value) in [
+            (rdf::type_, &cat::Sample.as_simple() as &dyn InsertIntoGraph),
+            (cat::role, &self.role.as_simple()),
+            (cat::vialShape, &self.vial_type.as_simple()),
+            (allores::AFR_0002464, &self.vial_type.as_simple()),
+            (cat::expectedDatum, &self.expected_datum),
+            (cat::hasSample, &self.has_sample),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: prop.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+        
+        // NOTE: for container_info, we attach triples directly to the sample
+        let _ = &self.container.insert_into(graph, iri.clone())?;
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct SampleItem {
     #[serde(rename = "sampleID")]
@@ -101,7 +283,35 @@ pub struct SampleItem {
     pub has_chemical: Chemical,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+impl InsertIntoGraph for SampleItem {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+
+        for (prop, value) in [
+            (rdf::type_, &cat::Sample.as_simple() as &dyn InsertIntoGraph),
+            (purl::identifier, &self.sample_id.as_simple()),
+            (cat::role, &self.role.as_simple()),
+            (cat::internalBarCode, &self.internal_bar_code.as_simple()),
+            (alloqual::AFQ_0000111, &self.physical_state.as_simple()),
+            (cat::expectedDatum, &self.expected_datum),
+            (cat::measuredQuantity, &self.measured_quantity),
+            (allores::AFR_0002036, &self.concentration),
+            (cat::hasChemical, &self.has_chemical),
+        ] {
+            value.attach_into(
+            graph,
+            Link {
+                source_iri: iri.clone(),
+                pred: prop.as_simple(),
+                target_iri: None,
+            },
+            )?;
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Chemical {
     #[serde(rename = "chemicalID")]
@@ -119,10 +329,96 @@ pub struct Chemical {
     pub density: Option<Observation>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+impl InsertIntoGraph for Chemical {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+
+        for (prop, value) in [
+            (rdf::type_, &obo::CHEBI_25367.as_simple() as &dyn InsertIntoGraph),
+            (purl::identifier, &self.chemical_id.as_simple()),
+            (cat::chemicalName, &self.chemical_name.as_simple()),
+            (allores::AFR_0001952, &self.molecular_formula.as_simple()),
+            (allores::AFR_0002295, &self.smiles.as_simple()),
+            (allores::AFR_0002294, &self.molecular_mass),
+            (allores::AFR_0002296, &self.inchi.as_simple()),
+            (cat::casNumber, &self.cas_number.as_ref().clone().map(|s| s.as_simple())),
+            (cat::swissCatNumber, &self.swiss_cat_number.as_ref().clone().map(|s| s.as_simple())),
+            (schema::keywords, &self.keywords.as_ref().clone().map(|s| s.as_simple())),
+            (obo::PATO_0001019, &self.density),
+        ] {
+            value.attach_into(
+            graph,
+            Link {
+                source_iri: iri.clone(),
+                pred: prop.as_simple(),
+                target_iri: None,
+            },
+            )?;
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct ContainerPositionQuantityItem {
     #[serde(rename = "containerID")]
     pub container_id: String,
     pub position: String,
     pub quantity: Observation,
+}
+
+impl InsertIntoGraph for ContainerPositionQuantityItem {
+    fn insert_into(&self, graph: &mut LightGraph, iri: SimpleTerm) -> anyhow::Result<()> {
+
+
+        for (pred, value) in [
+            (rdf::type_, &cat::ContainerPositionAndQuantity.as_simple() as &dyn InsertIntoGraph),
+            (cat::containerID, &self.container_id.as_simple()),
+            (allores::AFR_0002240, &self.position.as_simple()),
+            (qudt::quantity, &self.quantity),
+        ] {
+            value.attach_into(
+                graph,
+                Link {
+                    source_iri: iri.clone(),
+                    pred: pred.as_simple(),
+                    target_iri: None,
+                },
+            )?;
+        }
+
+        Ok(())
+    }
+}
+
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use sophia::iri::IriRef;
+    use sophia_api::term::Term;
+
+    use crate::{
+        graph::{graph_builder::GraphBuilder, insert_into::InsertIntoGraph},
+        models::{ErrorMargin, Observation},
+    };
+
+    #[test]
+    fn test_observation_to_triples() -> anyhow::Result<()> {
+        let observation = Observation {
+            value: 42.0,
+            unit: Unit::DegC,
+            error_margin: Some(ErrorMargin {
+                value: 0.5,
+                unit: Unit::DegC,
+            }),
+        };
+
+        let mut b = GraphBuilder::new();
+        let i = IriRef::new_unchecked("http://test.com/my-obersvation");
+        observation.insert_into(&mut b.graph, i.as_simple())?;
+        println!("Graph\n{}", b.serialize_to_turtle().unwrap());
+
+        Ok(())
+    }
 }
